TODO NUEVO

+++ ApptivoAPI/API/API/API/Models/LatLng.cs
@@ -7,7 +7,7 @@ namespace API.Models
 {
     public class LatLng
     {
-        public double Lat { get; set; }
-        public double Lng { get; set; }
+        public float Lat { get; set; }
+        public float Lng { get; set; }
     }
 }
